add support for dryrun via preparelocal
